ORM correction
added and removed some attributes.
replaced some entities.
moved dto into model package and created entity package.

+++ src/main/java/com/mentorship/javaeats/repository/AuditingRepository.java
@@ -1,7 +1,8 @@
 package com.mentorship.javaeats.repository;
 
-import com.mentorship.javaeats.model.Auditing;
+
+import com.mentorship.javaeats.model.Entity.Auditing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditingRepository extends JpaRepository<Auditing, Integer> {
+public interface AuditingRepository extends JpaRepository<Auditing, Long> {
 }
