+++ src/main/java/com/mentorship/javaeats/repository/OrderItemRepository.java
@@ -1,7 +1,8 @@
 package com.mentorship.javaeats.repository;
 
-import com.mentorship.javaeats.model.OrderItem;
+
+import com.mentorship.javaeats.model.Entity.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
 }
