+++ src/main/java/com/mentorship/javaeats/repository/TransactionStatusRepository.java
@@ -1,7 +1,8 @@
 package com.mentorship.javaeats.repository;
 
-import com.mentorship.javaeats.model.TransactionStatus;
+
+import com.mentorship.javaeats.model.Entity.TransactionStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionStatusRepository extends JpaRepository<TransactionStatus, Integer> {
+public interface TransactionStatusRepository extends JpaRepository<TransactionStatus, Long> {
 }
