+++ src/main/java/com/mentorship/javaeats/repository/UserTypeRepository.java
@@ -1,7 +1,8 @@
 package com.mentorship.javaeats.repository;
 
-import com.mentorship.javaeats.model.UserType;
+
+import com.mentorship.javaeats.model.Entity.UserType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserTypeRepository extends JpaRepository<UserType, Integer> {
+public interface UserTypeRepository extends JpaRepository<UserType, Long> {
 }
